Move nft related headers to protocol
